requirements.txt
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
-e git+https://github.com/SwiftAkira/KAI.git@732a23e6981b5dc23f98326f104d384939f8fc82#egg=pressure_agi
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
torch==2.7.1
typer==0.16.0
typing_extensions==4.14.0
